refactor: create utils file
