+++ src/components/Tooltip/Tooltip.tsx
@@ -3,16 +3,16 @@ import {
   Dimensions,
   View,
   LayoutChangeEvent,
-  LayoutRectangle,
   StyleSheet,
   StyleProp,
   ViewStyle,
 } from 'react-native';
 
-import { useInternalTheme } from '../core/theming';
-import { addEventListener } from '../utils/addEventListener';
-import Portal from './Portal/Portal';
-import Text from './Typography/Text';
+import { useInternalTheme } from '../../core/theming';
+import { addEventListener } from '../../utils/addEventListener';
+import Portal from '../Portal/Portal';
+import Text from '../Typography/Text';
+import { getTooltipPosition, Measurement } from './utils';
 
 export type Props = {
   /**
@@ -35,90 +35,6 @@ export type Props = {
    * Style for the wrapper view
    */
   wrapperStyle?: StyleProp<ViewStyle>;
-};
-
-type ChildrenMeasurement = {
-  width: number;
-  height: number;
-  pageX: number;
-  pageY: number;
-};
-
-type TooltipLayout = LayoutRectangle;
-
-type Measurement = {
-  children: ChildrenMeasurement;
-  tooltip: TooltipLayout;
-  measured: boolean;
-};
-
-/**
- * Return true when the tooltip center x-coordinate relative to the wrapped element is negative.
- * The tooltip will be placed at the starting x-coordinate from the wrapped element.
- */
-const overflowLeft = (center: number): boolean => {
-  return center < 0;
-};
-
-/**
- * Return true when the tooltip center x-coordinate + tooltip width is greater than the layout width
- * The tooltip width will grow from right to left relative to the wrapped element.
- */
-const overflowRight = (center: number, tooltipWidth: number): boolean => {
-  const { width: layoutWidth } = Dimensions.get('window');
-
-  return center + tooltipWidth > layoutWidth;
-};
-
-/**
- * Return true when the children y-coordinate + its height + tooltip height is greater than the layout height.
- * The tooltip will be placed at the top of the wrapped element.
- */
-const overflowBottom = (
-  childrenY: number,
-  childrenHeight: number,
-  tooltipHeight: number
-): boolean => {
-  const { height: layoutHeight } = Dimensions.get('window');
-
-  return childrenY + childrenHeight + tooltipHeight > layoutHeight;
-};
-
-const getTooltipXPosition = (
-  { pageX: childrenX, width: childrenWidth }: ChildrenMeasurement,
-  { width: tooltipWidth }: TooltipLayout
-): number => {
-  const center = childrenX + (childrenWidth - tooltipWidth) / 2;
-
-  if (overflowLeft(center)) return childrenX;
-
-  if (overflowRight(center, tooltipWidth))
-    return childrenX + childrenWidth - tooltipWidth;
-
-  return center;
-};
-
-const getTooltipYPosition = (
-  { pageY: childrenY, height: childrenHeight }: ChildrenMeasurement,
-  { height: tooltipHeight }: TooltipLayout
-): number => {
-  if (overflowBottom(childrenY, childrenHeight, tooltipHeight))
-    return childrenY - tooltipHeight;
-
-  return childrenY + childrenHeight;
-};
-
-const getTooltipPosition = ({
-  children,
-  tooltip,
-  measured,
-}: Measurement): {} | { left: number; top: number } => {
-  if (!measured) return {};
-
-  return {
-    left: getTooltipXPosition(children, tooltip),
-    top: getTooltipYPosition(children, tooltip),
-  };
 };
 
 /**
